calculator/requirements.txt
pandas==1.3.5
openpyxl==3.0.10
pyyaml==6.0
argparse==1.4.0
jinja2==3.1.2